+++ nequip/scripts/deploy.py
@@ -9,6 +9,7 @@ import argparse
 import pathlib
 import logging
 import yaml
+import itertools
 
 # This is a weird hack to avoid Intel MKL issues on the cluster when this is called as a subprocess of a process that has itself initialized PyTorch.
 # Since numpy gets imported later anyway for dataset stuff, this shouldn't affect performance.
@@ -129,7 +130,7 @@ def load_deployed_model(
 
 def main(args=None):
     parser = argparse.ArgumentParser(
-        description="Create and view information about deployed NequIP potentials."
+        description="Deploy and view information about previously deployed NequIP models."
     )
     # backward compat for 3.6
     if sys.version_info[1] > 6:
@@ -145,6 +146,11 @@ def main(args=None):
         "model_path",
         help="Path to a deployed model file.",
         type=pathlib.Path,
+    )
+    info_parser.add_argument(
+        "--print-config",
+        help="Print the full config of the model.",
+        action="store_true",
     )
 
     build_parser = subparsers.add_parser("build", help="Build a deployment model")
@@ -169,13 +175,25 @@ def main(args=None):
     logging.basicConfig(level=getattr(logging, args.verbose.upper()))
 
     if args.command == "info":
-        model, metadata = load_deployed_model(args.model_path, set_global_options=False)
-        del model
+        model, metadata = load_deployed_model(
+            args.model_path, set_global_options=False, freeze=False
+        )
         config = metadata.pop(CONFIG_KEY)
-        metadata_str = "\n".join("  %s: %s" % e for e in metadata.items())
-        logging.info(f"Loaded TorchScript model with metadata:\n{metadata_str}\n")
-        logging.info("Model was built with config:")
-        print(config)
+        if args.print_config:
+            print(config)
+        else:
+            metadata_str = "\n".join("  %s: %s" % e for e in metadata.items())
+            logging.info(f"Loaded TorchScript model with metadata:\n{metadata_str}\n")
+            logging.info(
+                f"Model has {sum(p.numel() for p in model.parameters())} weights"
+            )
+            logging.info(
+                f"Model has {sum(p.numel() for p in model.parameters() if p.requires_grad)} trainable weights"
+            )
+            logging.info(
+                f"Model weights and buffers take {sum(p.numel() * p.element_size() for p in itertools.chain(model.parameters(), model.buffers())) / (1024 * 1024):.2f} MB"
+            )
+            logging.debug(f"Model had config:\n{config}")
 
     elif args.command == "build":
         if args.model and args.train_dir: